app/requirements.txt
pymongo
Pillow